requirements.txt
Django==4.2.3
django_countries==7.5.1
Pillow==10.0.0
django-phonenumber-field[phonenumberslite]

